requirements.txt
pyyaml>=6.0
aiohttp>=3.8.0
Pillow>=9.0.0
